+++ src/pharmpy/tools/__init__.py
@@ -20,6 +20,7 @@ __all__ = (
     'run_tool',  # pyright: ignore [reportUnsupportedDunderAll]
     'summarize_individuals',  # pyright: ignore [reportUnsupportedDunderAll]
     'summarize_individuals_count_table',  # pyright: ignore [reportUnsupportedDunderAll]
+    'write_results',  # pyright: ignore [reportUnsupportedDunderAll]
 )
 
 
@@ -33,6 +34,7 @@ _run_keys = {
     'retrieve_final_model',
     'retrieve_models',
     'run_tool',
+    'write_results',
 }
 
 _func_keys = {