+++ CHANGELOG.rst
@@ -40,6 +40,7 @@ Changes
 * Move predict_* functions from modeling to pharmpy.tools
 * Move summarize_individuals and summarize_individuals_count_table to pharmpy.tools
 * Move print_fit_summary to pharmpy.tools
+* Move write_results to pharmpy.tools
 
 Bug fixes
 =========